Refactor Server class to extract metacom
Refs: https://github.com/metarhia/impress/issues/1236
