+++ lib/worker.js
@@ -6,7 +6,7 @@ const application = require('./application.js');
 const Config = require('@metarhia/config');
 const metalog = require('metalog');
 const Database = require('./database.js');
-const Server = require('./server.js');
+const Metacom = require('./metacom.js');
 const initAuth = require('./auth.js');
 const Console = require('./console.js');
 
@@ -29,7 +29,7 @@ const Console = require('./console.js');
     if (worker.threadId === 1) logger.error('Can not load TLS certificates');
   }
   application.db = new Database(config.sections.database);
-  application.server = new Server(config.sections.server);
+  application.server = new Metacom.Server(config.sections.server);
   application.auth = initAuth();
   application.sandboxInject('auth', application.auth);
   await application.init();
